cleaning old dde methods
