+++ control/tests/delay_lti_test.py
@@ -180,96 +180,3 @@ class TestDelayLtiMethods:
         dlti_cat = cat_func(dlti1, dlti2)
         assert np.allclose(dlti_cat.P.A, np.array(expected_A))
         assert np.allclose(dlti_cat.tau, np.array(expected_tau))
-
-
-class TestTimeResp:
-    def test_siso_delayed_step_response(self, delay_siso_tf, simple_siso_tf):
-        from control.timeresp import step_response
-        timepts = np.linspace(0, 10, 1001)
-        step = step_response(simple_siso_tf, timepts=timepts)
-        delay_step = step_response(delay_siso_tf, timepts=timepts)
-        # Construct a manually delayed step response by shifting the step response
-        hand_delayed_step = np.zeros_like(step.y[0][0])
-        count = 0
-        for i, t in enumerate(step.t):
-            if t >= 1.5:
-                hand_delayed_step[i] = step.y[0][0][count]
-                count += 1
-        plt.figure()
-        plt.plot(delay_step.y[0][0] - hand_delayed_step)
-        plt.legend()
-        plt.show()
-        assert np.allclose(delay_step.y[0][0], hand_delayed_step)
-
-    def test_siso_delayed_step_response_mos(self, delay_siso_tf, simple_siso_tf):
-        from control.timeresp import step_response
-        timepts = np.linspace(0, 10, 1001)
-        step = step_response(simple_siso_tf, timepts=timepts)
-        delay_step = step_response(delay_siso_tf, timepts=timepts, use_mos=True)
-        # Construct a manually delayed step response by shifting the step response
-        hand_delayed_step = np.zeros_like(step.y[0][0])
-        count = 0
-        for i, t in enumerate(step.t):
-            if t >= 1.5:
-                hand_delayed_step[i] = step.y[0][0][count]
-                count += 1
-
-        plt.figure()
-        plt.plot(delay_step.y[0][0] - hand_delayed_step)
-        plt.legend()
-        plt.show()
-        assert np.allclose(delay_step.y[0][0], hand_delayed_step, atol=1e-5)
-
-    # wood berry step response compared to julia
-    def test_mimo_step_response(self, wood_berry, plot=False):
-        from control.timeresp import step_response
-        import matplotlib.pyplot as plt
-        timepts = np.linspace(0, 100, 10001)
-        step = step_response(wood_berry, timepts=timepts, use_mos=True)
-        #step_lsoda = step_response(wood_berry, timepts=timepts)
-
-        if plot:
-            plt.figure()
-            plt.plot(step.y[0][0] - julia_json["TestTimeResp"]["test_mimo_step_response"]["y11"])
-            plt.plot(step.y[1][0] - julia_json["TestTimeResp"]["test_mimo_step_response"]["y21"])
-            plt.plot(step.y[0][1] - julia_json["TestTimeResp"]["test_mimo_step_response"]["y12"])
-            plt.plot(step.y[1][1] - julia_json["TestTimeResp"]["test_mimo_step_response"]["y22"])
-            plt.show()
-
-        assert np.allclose(step.y[0][0], julia_json["TestTimeResp"]["test_mimo_step_response"]["y11"], atol=1e-5)
-        assert np.allclose(step.y[0][1], julia_json["TestTimeResp"]["test_mimo_step_response"]["y12"], atol=1e-5)
-        assert np.allclose(step.y[1][1], julia_json["TestTimeResp"]["test_mimo_step_response"]["y22"], atol=1e-5)
-        assert np.allclose(step.y[1][0], julia_json["TestTimeResp"]["test_mimo_step_response"]["y21"], atol=1e-5)
-        
-
-    def test_forced_response(self, delay_siso_tf, simple_siso_tf, plot=False):
-        from control.timeresp import forced_response
-
-        timepts = np.linspace(0, 10, 1001)    
-        inputs = np.sin(timepts)
-        resp = forced_response(simple_siso_tf, timepts=timepts, inputs=inputs)
-        delay_resp = forced_response(delay_siso_tf, timepts=timepts, inputs=inputs, use_mos=True)
-        hand_delayed_resp = np.zeros_like(resp.y[0])
-        count = 0
-        for i, t in enumerate(resp.t):
-            if t >= 1.5:
-                hand_delayed_resp[i] = resp.y[0][count]
-                count += 1
-
-        # Optionally, inspect the plot:
-        plot=True
-        if plot:
-            plt.figure()
-            plt.plot(resp.t, inputs, label="input")
-            plt.plot(resp.t, resp.y[0], label="response")
-            plt.plot(resp.t, delay_resp.y[0], label="delay LTI smaller step size")
-            #plt.plot(delay_resp_few.t, delay_resp_few.y[0], label="delay LTI bigger step size")
-            plt.plot(resp.t, hand_delayed_resp, label="hand delay")
-            plt.legend()
-            plt.show()
-
-            plt.figure()
-            plt.plot(resp.t, delay_resp.y[0] - hand_delayed_resp)
-            plt.show()
-
-        assert np.allclose(delay_resp.y[0], hand_delayed_resp, atol=1e-5)
